🐛 Fix black modal & deleting wrong feedback
